Add return message to ams for ssp
