+++ bin/Untitled-1.json
@@ -1,0 +1,18 @@
+{
+    "host": "msg-devel.argo.grnet.gr",
+    "project" : "rciam-service-registry",
+    "pull_topic" : "prod_egi_service_mitreid",
+    "token" : "a7eee6262c719a466273e1b97a9946bbb980c694",
+    "pull_sub" : "prod_egi_service_oidc_mitreid_1",
+    "agent_id": "1",
+    "pub_topic": "prod_deployment_response",
+    "metadata": "/path/to/ssp/metadata/file.php",
+    "metadata_key": "ssp-metadata-key",
+    "ssp_url": "http://localhost/proxy/module.php/cron/cron.php",
+    "log_conf": "conf/logger.conf",
+    "mitreid_url": "https://aai-dev.egi.eu/oidc" ,
+    "mitreid_refresh_token": "eyJhbGciOiJub25lIn0.eyJleHAiOjE2MzgwODk1MDAsImp0aSI6ImQ5YTU3YjI0LWU1NDItNDAwZS1iMjZiLWMyMTk3YWQ5NWY0NiJ9.",
+    "mitreid_client_id": "7ccaaeec-98da-452c-b2d4-f833165f2a02",
+    "mitreid_client_secret": "AJkvmmTvFjBs38Q_qTV-7TaYuJqH6A7UT_9At02u6sFQNlK2gscc4UqvGtGsOApqBgoD_bshVzQNjEO8e7q9LJg",
+    "time_interval": 5
+  }
