--- bin/common.py
@@ -1,11 +1,0 @@
-import logging.config
-import logging
-
-def get_log_conf(log_config_file=None):
-    """
-    Method that searches and gets the default location of configuration and logging configuration
-    """
-    if log_config_file is not None:
-        logging.config.fileConfig(log_config_file, disable_existing_loggers=False)
-    else:
-        logging.basicConfig(level=logging.INFO, format=logging.BASIC_FORMAT)
